requirements.txt
python-telegram-bot==22.0
boto3==1.37.23
dotenv==0.9.9
anthropic==0.49.0